Fixed ping mechanism not working
