+++ changelog.d/1676.fixed.md
@@ -1,0 +1,1 @@
+Fixed ping mechanism not working
