+++ mirrord/cli/Cargo.toml
@@ -23,6 +23,7 @@ mirrord-kube = { path = "../kube" }
 mirrord-config = { path = "../config" }
 mirrord-protocol = { path = "../protocol" }
 mirrord-console = { path = "../console" }
+mirrord-connection = { path = "../connection" }
 mirrord-analytics = { path = "../analytics" }
 
 actix-codec.workspace = true
