+++ mirrord/connection/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "mirrord-connection"
+version.workspace = true
+authors.workspace = true
+description.workspace = true
+documentation.workspace = true
+readme.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+keywords.workspace = true
+categories.workspace = true
+publish.workspace = true
+edition.workspace = true
+
+[dependencies]
+mirrord-protocol = { path = "../protocol" }
+
+tokio.workspace = true
+tracing.workspace = true
+actix-codec.workspace = true
+futures.workspace = true
