+++ mirrord/layer/src/connection.rs
@@ -3,18 +3,15 @@
 //! The layer will connect to the internal proxy to communicate with the agent.
 use std::net::SocketAddr;
 
-use futures::{SinkExt, StreamExt};
-use mirrord_protocol::{ClientCodec, ClientMessage, DaemonMessage, LogLevel};
+use mirrord_protocol::{ClientMessage, DaemonMessage};
+use mirrord_connection::wrap_raw_connection;
 use tokio::{
     net::TcpStream,
-    sync::mpsc::{self, Receiver, Sender},
+    sync::mpsc::{Receiver, Sender},
 };
-use tracing::{error, info, trace, warn};
+use tracing::error;
 
 use crate::graceful_exit;
-
-/// Size of the channel used in [`wrap_raw_connection`] for [`ClientMessage`]s.
-const CONNECTION_CHANNEL_SIZE: usize = 1000;
 
 const FAIL_STILL_STUCK: &str = r#"
 - If you're still stuck:
@@ -41,70 +38,4 @@ pub(crate) async fn connect_to_proxy(
         }
     };
     wrap_raw_connection(stream)
-}
-
-/// Creates the task that handles the messaging between layer/agent.
-/// It does the encoding/decoding of protocol.
-fn wrap_raw_connection(
-    stream: TcpStream,
-) -> (mpsc::Sender<ClientMessage>, mpsc::Receiver<DaemonMessage>) {
-    let mut codec = actix_codec::Framed::new(stream, ClientCodec::new());
-
-    let (client_in_tx, mut client_in_rx) = mpsc::channel(CONNECTION_CHANNEL_SIZE);
-    let (daemon_out_tx, daemon_out_rx) = mpsc::channel(CONNECTION_CHANNEL_SIZE);
-
-    tokio::spawn(async move {
-        loop {
-            tokio::select! {
-                msg = client_in_rx.recv() => {
-                    match msg {
-                        Some(msg) => {
-                            trace!("Sending client message: {:?}", &msg);
-                            if let Err(fail) = codec.send(msg).await {
-                                error!("Error sending client message: {:#?}", fail);
-                                break;
-                            }
-                        }
-                        None => {
-                            info!("mirrord-kube: initiated disconnect from agent");
-
-                            break;
-                        }
-                    }
-                }
-                daemon_message = codec.next() => {
-                    match daemon_message {
-                        Some(Ok(DaemonMessage::LogMessage(log_message))) => {
-                            match log_message.level {
-                                LogLevel::Warn => {
-                                    warn!(message = log_message.message, "Daemon sent log message")
-                                }
-                                LogLevel::Error => {
-                                    error!(message = log_message.message, "Daemon sent log message")
-                                }
-                            }
-                        },
-                        Some(Ok(msg)) => {
-                            if let Err(fail) = daemon_out_tx.send(msg).await {
-                                error!("DaemonMessage dropped: {:#?}", fail);
-
-                                break;
-                            }
-                        }
-                        Some(Err(err)) => {
-                            error!("Error receiving daemon message: {:?}", err);
-                            break;
-                        }
-                        None => {
-                            error!("agent disconnected");
-
-                            break;
-                        }
-                    }
-                }
-            }
-        }
-    });
-
-    (client_in_tx, daemon_out_rx)
 }
